New update: 11:18PM

+++ server/bread/routes/cart/sargents/delete_cart_item.js
@@ -9,11 +9,16 @@ deleteCartItemRoute.post(endpoints.DELETE_CART_ITEM, async (req, res) => {
     console.log('They tried to delete me!');
     try {
         const { ownerId, itemId } = req.body;
+
+        if (!ownerId || !itemId) {
+            throw new Error('Some passed arguments are undefined (ownerId, itemId)');
+        }
+
         let cart = await arsenal.findCart(ownerId);
         let updatedCart = await arsenal.deleteCartItem(cart, itemId);
-        await updatedCart.save();
-        console.log(updatedCart);
 
+        console.log(updatedCart);
+        await updatedCart.save();
         res.status(200).json(updatedCart);
 
     } catch (e) {
